Feat/be/cloud (#24)
* fix: 배포 스크립트 오류 수정

* refactor: 소셜 로그인 시 회원가입 로직 수정

* refactor: JWT 액세스 토큰 발급 시 Bearer 붙이기 개선

* refactor: Security 권한 적용된 테스트 코드 수정

* bulid: 빌드 생성 시각 추가 및 스크립트 수정

* bulid: 빌드 생성 시각 추가 및 스크립트 수정 - 2

* feat: Member 테스트 코드 작성

* feat: JwtTokenizer 테스트 코드 작성

+++ server/src/main/java/com/lookatme/server/member/controller/MemberController.java
@@ -1,6 +1,9 @@
 package com.lookatme.server.member.controller;
 
+import com.lookatme.server.auth.dto.MemberPrincipal;
 import com.lookatme.server.common.dto.MultiResponseDto;
+import com.lookatme.server.exception.ErrorCode;
+import com.lookatme.server.exception.ErrorLogicException;
 import com.lookatme.server.member.dto.MemberDto;
 import com.lookatme.server.member.entity.Member;
 import com.lookatme.server.member.mapper.MemberMapper;
@@ -9,6 +12,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -47,7 +51,12 @@ public class MemberController {
 
     @PatchMapping("/{memberId}")
     public ResponseEntity<?> updateMember(@Positive @PathVariable long memberId,
-                                          @Valid @RequestBody MemberDto.Patch patchDto) {
+                                          @Valid @RequestBody MemberDto.Patch patchDto,
+                                          @AuthenticationPrincipal MemberPrincipal memberPrincipal) {
+        // 로그인 유저와 정보가 다르다면 수정 불가 (접근 권한 X)
+        if (memberPrincipal.getMemberId() != memberId) {
+            throw new ErrorLogicException(ErrorCode.UNAUTHORIZED);
+        }
         Member member = mapper.memberPatchDtoToMember(patchDto, memberId);
         return new ResponseEntity<>(
                 mapper.memberToMemberResponse(memberService.updateMember(member)),
@@ -55,7 +64,12 @@ public class MemberController {
     }
 
     @DeleteMapping("/{memberId}")
-    public ResponseEntity<?> deleteMember(@Positive @PathVariable long memberId) {
+    public ResponseEntity<?> deleteMember(@Positive @PathVariable long memberId,
+                                          @AuthenticationPrincipal MemberPrincipal memberPrincipal) {
+        // 로그인 유저와 정보가 다르다면 수정 불가 (접근 권한 X)
+        if (memberPrincipal.getMemberId() != memberId) {
+            throw new ErrorLogicException(ErrorCode.UNAUTHORIZED);
+        }
         memberService.deleteMember(memberId);
         return new ResponseEntity<>("회원탈퇴 되었습니다", HttpStatus.NO_CONTENT);
     }
